Move controllers package into pkg dir
